[CCUBE-1041][SR] Work in progress; Refactoring
