--- src/components/panel/panel.tsx
@@ -1,9 +1,0 @@
-import { SidePanel } from "./side-panel";
-
-export const Panel = () => {
-    return (
-        <SidePanel>
-            <h2>Hello World!</h2>
-        </SidePanel>
-    );
-};
