+++ src/form-builder.tsx
@@ -1,0 +1,10 @@
+import { Main } from "./components";
+import { BuilderProvider } from "./context-providers/builder";
+
+export const FormBuilder = () => {
+    return (
+        <BuilderProvider>
+            <Main>Hello</Mian
+        </BuilderProvider>
+    );
+};
