+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from "./main";
+export * from "./form-builder";
